Issue #31 related - Commenting feature added for offers
The Comment model is updated. Now, the Comment model is a parent class.
There are 3 subclasses; OfferComment, RequestComment and
HandshakeComment. All of these classes are marked as entities,
including the Comment model itself. Comment model is marked as an
entity; because the Play Framework supports model inheritance as in
"concrete parent classes".