Issue #28 related - Private messaging feature is implemented
The backbones of the private messaging feature is implemented.
Currently, a user can send a private message to another user (including
himself, which in fact should be avoided) from his/her profile page.
There are inbox and outbox pages, which list received and sent messages
respectively. The front-end design should be modified. Also, there is a
new field under the user menu (which can be opened by clicking on the
logged-in user's name at the top-right hand side) that shows a link to
the inbox including the unread messages count.